# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para fornecer sugestões de endereço e outros serviços relacionados",
        "title": "UrbVerde BFF",
        "contact": {
            "name": "API Support",
            "url": "http://www.urbverde.com.br/",
            "email": "comunica.urbverde@usp.br"
        },
        "license": {
            "name": "???",
            "url": "???"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/cards/square/inequality": {
            "get": {
                "description": "Retorna dados de desigualdade para a camada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/square"
                ],
                "summary": "Retorna dados sobre desigualdade",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/square/info": {
            "get": {
                "description": "Retorna dados adicionais para a camada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/square"
                ],
                "summary": "Retorna dados adicionais para a camada de praças e parques",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/square/parks": {
            "get": {
                "description": "Retorna dados de parques e praças para a camada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/square"
                ],
                "summary": "Retorna dados dos parques e praças",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/square/ranking": {
            "get": {
                "description": "Retorna dados para a construção do ranking de praças e parques",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/square"
                ],
                "summary": "Retorna dados de ranking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ano dos dados",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.RankingData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/temperature/heat": {
            "get": {
                "description": "Retorna os dados de calor extremo para o município e ano fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/temperature"
                ],
                "summary": "Retorna dados de calor extremo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ano dos dados",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/temperature/info": {
            "get": {
                "description": "Retorna dados adicionais para o município fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/temperature"
                ],
                "summary": "Retorna dados adicionais",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/temperature/ranking": {
            "get": {
                "description": "Retorna os dados de ranking em clima para o município e ano fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/temperature"
                ],
                "summary": "Retorna dados de ranking de clima",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ano dos dados",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.RankingData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/temperature/weather": {
            "get": {
                "description": "Retorna os dados de temperatura para o município e ano fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/temperature"
                ],
                "summary": "Retorna dados de temperatura",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ano dos dados",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/vegetation/cover": {
            "get": {
                "description": "Retorna dados relacionados à cobertura vegetal para o município fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/vegetation"
                ],
                "summary": "Retorna dados relacionados à cobertura vegetal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ano dos dados",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/vegetation/inequality": {
            "get": {
                "description": "Retorna dados relacionados à desigualdade ambiental e a vegetação para o município fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/vegetation"
                ],
                "summary": "Retorna dados relacionados à desigualdade ambiental e a vegetação",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ano dos dados",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/vegetation/info": {
            "get": {
                "description": "Retorna dados adicionais para a camada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/vegetation"
                ],
                "summary": "Retorna dados adicionais para a vegetação",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardsDataItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/vegetation/ranking": {
            "get": {
                "description": "Retorna dados para a construção do ranking de desigualdade ambiental e a vegetação",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards/vegetation"
                ],
                "summary": "Retorna dados de ranking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de município",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ano dos dados",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.RankingData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cards.CardsDataItem": {
            "type": "object",
            "properties": {
                "subtitle": {
                    "type": "string",
                    "example": "Abaixo da média nacional de 0"
                },
                "title": {
                    "type": "string",
                    "example": "Nível de ilha de calor"
                },
                "value": {
                    "type": "string",
                    "example": "25°C"
                }
            }
        },
        "cards.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Erro ao processar a solicitação"
                }
            }
        },
        "cards.RankingData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cards.RankingDataItem"
                    }
                },
                "subtitle": {
                    "type": "string",
                    "example": "Posição do seu município entre os 645 do Estado de São Paulo"
                },
                "title": {
                    "type": "string",
                    "example": "Municipios do Estado"
                }
            }
        },
        "cards.RankingDataItem": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "integer"
                },
                "of": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}